+++ src/PhpSpreadsheet/Writer/Pdf/Mpdf.php
@@ -40,6 +40,13 @@ class Mpdf extends Pdf
 
         //  Create PDF
         $config = ['tempDir' => $this->tempDir . '/mpdf'];
+        $restoreHandler = false;
+        if (PHP_VERSION_ID >= self::$temporaryVersionCheck) {
+            // @codeCoverageIgnoreStart
+            set_error_handler(self::specialErrorHandler(...));
+            $restoreHandler = true;
+            // @codeCoverageIgnoreEnd
+        }
         $pdf = $this->createExternalWriterInstance($config);
         $ortmp = $orientation;
         $pdf->_setPageSize($paperSize, $ortmp);
@@ -83,7 +90,28 @@ class Mpdf extends Pdf
         $str = $pdf->Output('', 'S');
         fwrite($fileHandle, $str);
 
+        if ($restoreHandler) {
+            restore_error_handler(); // @codeCoverageIgnore
+        }
         parent::restoreStateAfterSave();
+    }
+
+    protected static int $temporaryVersionCheck = 80500;
+
+    /**
+     * Temporary handler for Php8.5 waiting for Dompdf release.
+     *
+     * @codeCoverageIgnore
+     */
+    public function specialErrorHandler(int $errno, string $errstr, string $filename, int $lineno): bool
+    {
+        if ($errno === E_DEPRECATED) {
+            if (preg_match('/Providing an empty string is deprecated/', $errstr) === 1) {
+                return true;
+            }
+        }
+
+        return false; // continue error handling
     }
 
     /**