+++ tests/PhpSpreadsheetTests/Reader/Xlsx/XlsxTest.php
@@ -14,6 +14,7 @@ use PhpOffice\PhpSpreadsheet\Style\Conditional;
 use PhpOffice\PhpSpreadsheet\Style\Style;
 use PhpOffice\PhpSpreadsheet\Worksheet\AutoFilter;
 use PhpOffice\PhpSpreadsheet\Worksheet\PageSetup;
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
 class XlsxTest extends TestCase
@@ -194,11 +195,11 @@ class XlsxTest extends TestCase
         $reader = new Xlsx();
         $reader->setReadFilter(new OddColumnReadFilter());
         $spreadsheet = $reader->load($filename);
-        $data = $spreadsheet->getActiveSheet()->toArray();
-        $ref = [1.0, null, 3.0, null, 5.0, null, 7.0, null, 9.0, null];
+        $data = $spreadsheet->getActiveSheet()->toArray(formatData: false);
+        $ref = [1, null, 3, null, 5, null, 7, null, 9, null];
 
         for ($i = 0; $i < 10; ++$i) {
-            self::assertEquals($ref, \array_slice($data[$i], 0, 10, true));
+            self::assertSame($ref, \array_slice($data[$i], 0, 10, true));
         }
         $spreadsheet->disconnectWorksheets();
     }
@@ -237,7 +238,7 @@ class XlsxTest extends TestCase
      * Test if all whitespace is removed from a style definition string.
      * This is needed to parse it into properties with the correct keys.
      */
-    #[\PHPUnit\Framework\Attributes\DataProvider('providerStripsWhiteSpaceFromStyleString')]
+    #[DataProvider('providerStripsWhiteSpaceFromStyleString')]
     public function testStripsWhiteSpaceFromStyleString(string $string): void
     {
         $string = Xlsx::stripWhiteSpaceFromStyleString($string);